move token factory function to cassandra-protocol crate & add tests
